Created a url parser that dynamically updates the main app URL hash with the view type and options for those views.  Currently only committed the map view options, which are zoom and focus, but those work both forwards and backwards, i.e. you can enter them into the hash and it will load your selection AND it will dynamically update the URL as you move around the map.
If a URL is supplied that is missing a zoom level, that defaults to 12.  If it's missing a focus, it uses geoPlugin (http://www.geoplugin.com/) to look up the client's URL and sets that as the focus.  If both are missing, it does a fitBounds call.

I will work on the list view portions when I get more done.  That will at least accept the sort options.