Removed jsonpickle from manual.txt, as it is not used, and added python-dateutil, which is required for parsing date strings in mongo.
Added a default last_login field for the User test fixture as that is a NOT NULL field and sqlite will error out if it sees a null in that situation.

Fixed conflicting default and auto_now_add parameters for date_joined, which Django 1.8 now errors on as those are mutually exclusive.