Fixed #81 -- Changed target of static file compilation.
In development, Django cannot serve static files from the directory into which it places static files with collectstatic.  The LESS/CSS files and nunjucks templates were being compiled directly into /build, so the Gruntfile was changed to compile them into /artifacts.  The base settings.__init__ file now includes /artifacts in STATICFILES_DIRS so that the files in there can be served in development testing.

Also removed an import statement from core URLs that had mistakenly been brought over from the testing of ElasticSearch and which would cause Django to error out since the view it referenced does not exist in the core code.