Fixed #92 -- Moved test cases to a more standard location.
The test cases that were stored in core modules (those that would be
part of Ografy if it were ever released as a library) were moved to a
top level /tests directory.

These tests no longer work, but the libraries they test are
changing anyway. The code is left in and relocated rather than
deleted it so there is a reference moving forward.