Fixed #80 -- Add support for location estimation.
This codebase primarily adds location estimation. An internal endpoint
is called by the Event Post process if the Event does not have valid
coordinates. The endpoint searches the ElasticSearch store of Locations
for the user for the immediately preceding and/or following Locations in
time. Based on a user setting, the location will be set to the previous
location, next location, whichever one is closer in time, or a weighted
distance between the two.

Code has been written in a test view to get the user's coordinates and
create a new Location in the database with them at a regular interval.
The goal is to create a much broader set of Locations with which to
estimate the locations of Events. This will be merged into the main page
at a later date.

A new Location Settings page has been added to the user settings pages.
This page allows the user turn on or off the location posting described
in the previous paragraph. It also allows them to switch their location
estimation preference. Finally, it has button that will tell the server
to re-estimate all of their estimated locations. This is only active if
the most recent re-estimation was more than 5 days ago to prevent
excessive server load.