Fixed #207 -- Resolved connection association failures.
The custom pipeline step written for the PSA workflow did not account
for existing connections to a provider. This was causing most of the
server-side errors people were seeing with this process.

The process might still "fail" if it finds more than one unassociated
connection, but this is normal and is supposed to send the user
back to the Providers page; if they attempt to connect to that provider
again, it should work.

There may still be some lingering issues with this workflow, but the
application at least passes cursory tests. A more robust solution may
be required in the future.