Fixed #238 -- Implemented Saved Searches.
The system now tracks every user's searches. Whenever a user performs
a search, the JSON serialized filters and query text are saved to
ElasticSearch along with a hash of the search that is calculated
by repeatably sorting the filters such that any arrangement of the
exact same filters will produce the same hash. If a search with that
hash already exists, then the existing copy is updated rather than
creating a new one. Along with the serialized filters, query text,
and hash, we also save when the search was last run, any named filters,
and how many times that search has been run.

The user can also 'favorite' any search (terminology may change).
This process involves the user associating a name and/or icon
with that search. The user can select a color for the icon as well.
The user can delete the favorite status for a search, but currently
cannot delete a search from their history outright.

When a user is creating a search, the client queries the server
to see if that search has been run before. If it has, then any named
filters and favorite status will propogate to the client.

The user home page has been revamped to act as a place to manage
saved searches. The user can filter their saved searches by time and
how many times they have been run, as well as viewing a list of
only favorited searches. From here, they can edit the name and icon
of favorited searches as well as un-favorite favorited searches.
Clicking on a saved search will take the user to the explorer and
run that search.