Removing redundant popstate listener.
After the results preview was removed, the popstate listener that had
previously transitioned between the active facet and the results preview
was still in place. Furthermore the listener was causing issues with
Safari since Safari fires a popstate event on page refresh (but not
address bar enter key). This listener could have been fixed by checking
the current state, thus ignoring the initial Safari popstate, but it was
instead simply removed since it was no longer necessary anyway.

Resolves:
  * APP-79 -- Safari throws errors when refreshing the page.